chore: Restructured the historical page folder
