+++ frontend/client/package.json
@@ -3,64 +3,65 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
-      "@emotion/react": "^11.10.5",
-      "@emotion/styled": "^11.10.5",
-      "@hookform/resolvers": "^2.9.10",
-      "@mui/base": "^5.0.0-alpha.106",
-      "@mui/icons-material": "^5.10.14",
-      "@mui/material": "^5.10.14",
-      "@mui/styled-engine-sc": "^5.10.14",
-      "@reduxjs/toolkit": "^1.9.0",
-      "@testing-library/jest-dom": "^5.16.5",
-      "@testing-library/react": "^13.4.0",
-      "@testing-library/user-event": "^13.5.0",
-      "iconsax-react": "^0.0.8",
-      "react": "^18.2.0",
-      "react-dom": "^18.2.0",
-      "react-hook-form": "^7.39.4",
-      "react-icons": "^4.6.0",
-      "react-router-dom": "^6.4.3",
-      "react-router-hash-link": "^2.4.3",
-      "react-scripts": "5.0.1",
-      "redux": "^4.2.0",
-      "styled-components": "^5.3.6",
-      "web-vitals": "^2.1.4",
-      "yup": "^0.32.11"
+        "@emotion/react": "^11.10.5",
+        "@emotion/styled": "^11.10.5",
+        "@fontsource/inter": "^4.5.14",
+        "@hookform/resolvers": "^2.9.10",
+        "@mui/base": "^5.0.0-alpha.106",
+        "@mui/icons-material": "^5.10.14",
+        "@mui/material": "^5.10.14",
+        "@mui/styled-engine-sc": "^5.10.14",
+        "@reduxjs/toolkit": "^1.9.0",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^13.5.0",
+        "iconsax-react": "^0.0.8",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-hook-form": "^7.39.4",
+        "react-icons": "^4.6.0",
+        "react-router-dom": "^6.4.3",
+        "react-router-hash-link": "^2.4.3",
+        "react-scripts": "5.0.1",
+        "redux": "^4.2.0",
+        "styled-components": "^5.3.6",
+        "web-vitals": "^2.1.4",
+        "yup": "^0.32.11"
     },
     "scripts": {
-      "start": "react-scripts start",
-      "build": "CI=false && react-scripts build",
-      "test2": "react-scripts test",
-      "eject": "react-scripts eject",
-      "prepare": "cd ../.. && husky install frontend/client/.husky",
-      "test": "jest"
+        "start": "react-scripts start",
+        "build": "CI=false && react-scripts build",
+        "test2": "react-scripts test",
+        "eject": "react-scripts eject",
+        "prepare": "cd ../.. && husky install frontend/client/.husky",
+        "test": "jest"
     },
     "eslintConfig": {
-      "extends": [
-        "react-app",
-        "react-app/jest"
-      ]
+        "extends": [
+            "react-app",
+            "react-app/jest"
+        ]
     },
     "browserslist": {
-      "production": [
-        ">0.2%",
-        "not dead",
-        "not op_mini all"
-      ],
-      "development": [
-        "last 1 chrome version",
-        "last 1 firefox version",
-        "last 1 safari version"
-      ]
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
     },
     "devDependencies": {
-      "eslint": "^8.27.0",
-      "husky": "^8.0.2",
-      "jest": "^27.5.1",
-      "lint-staged": "^13.0.3",
-      "prettier": "2.7.1"
+        "eslint": "^8.27.0",
+        "husky": "^8.0.2",
+        "jest": "^27.5.1",
+        "lint-staged": "^13.0.3",
+        "prettier": "2.7.1"
     },
     "lint-staged": {
-      "**/*": "prettier --write --ignore-unknown"
+        "**/*": "prettier --write --ignore-unknown"
     }
-  }
+}
