+++ frontend/client/src/Pages/HistoricalCurrencyExchange/assets/index.js
@@ -1,0 +1,15 @@
+export {default as Arrow} from "./png/arrow.png"
+export {default as Flag} from "./png/Flag.png"
+export {default as Flag1} from "./png/Flag1.png"
+export {default as Flag2} from "./png/Flag2.png"
+export {default as Flag3} from "./png/Flag3.png"
+export {default as Flag4} from "./png/Flag4.png"
+export {default as Flag5} from "./png/Flag5.png"
+export {default as Flag6} from "./png/Flag6.png"
+export {default as Flag7} from "./png/Flag7.png"
+export {default as Flag8} from "./png/Flag8.png"
+export {default as Frame} from "./png/frame.png"
+export {default as Rectangle} from "./png/Rectangle 2.png"
+export {default as Rectangle1} from "./png/Rectangle 3.png"
+export {default as Vector} from "./png/Vector.png"
+
