+++ frontend/client/src/Pages/HistoricalCurrencyExchange/assets/png/Flag2.png
