+++ frontend/client/src/Pages/HistoricalCurrencyExchange/assets/png/arrow.png
