+++ frontend/client/src/Pages/HistoricalCurrencyExchange/assets/png/frame.png
