+++ frontend/client/src/Pages/HistoricalCurrencyExchange/components/Top.jsx
@@ -1,0 +1,13 @@
+import {TopComponentWrapper} from "../syles/HistoricalCurrencyExchangeStyles";
+import Converter from "./Converter"
+
+const TopComponent = () => {
+    return (
+      <TopComponentWrapper>
+        {/* <Calender /> */}
+        <Converter />
+      </TopComponentWrapper>
+    );
+  };
+
+export default TopComponent
